modify auctions stack + copy calendar page to index and remove calendar
